Use an img fallback for print
Always add an img fallback, so that we can use it when we print, since
background images are usually stripped from the print by user-agents.

The img element is initially hidden, so that we don't have multiple
elements racing for first draw, and potentially causing visual
'glitches'. It is only unhidden on browsers that do not support
background-size, or when printing.